requirements.txt
ddgs
crewai
pydantic
crewai_tools
python-dotenv
langchain_groq
duckduckgo-search
